Reject in catch instead of resolve
Initially, and I still prefer the UX, I set it up to notify with
details in Atom on failure and then resolve with null. The reason
for this is the necessary details for the error would be available to
the user, but would not cause the `atom-linter` to also freak out.

That being said it's failing in CI and no info can be had there, so I'm
rejecting instead. I started to check env.CI but that made me feel dirty